Re-factor app into separate pages
Also migrates the app to Typescript.
