+++ src/pages/forms.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+import {
+} from '../lib';
+
+const page = (<>
+  <h1>Welcome to HOF2!</h1>
+</>);
+
+export default page;
